+++ soundlib/plugins/dmo/WavesReverb.cpp
@@ -144,12 +144,12 @@ void WavesReverb::Resume()
 {
 	m_isResumed = true;
 	// Recalculate delays
-	uint32 delay0 = mpt::saturate_round<uint32>(m_SndFile.GetSampleRate() * 0.045f);
-	uint32 delay1 = mpt::saturate_round<uint32>(delay0 * 1.18920707f);	// 2^0.25
-	uint32 delay2 = mpt::saturate_round<uint32>(delay1 * 1.18920707f);
-	uint32 delay3 = mpt::saturate_round<uint32>(delay2 * 1.18920707f);
-	uint32 delay4 = mpt::saturate_round<uint32>((delay0 + delay2) * 0.11546667f);
-	uint32 delay5 = mpt::saturate_round<uint32>((delay1 + delay3) * 0.11546667f);
+	uint32 delay0 = mpt::saturate_round<uint32>(static_cast<float>(m_SndFile.GetSampleRate()) * 0.045f);
+	uint32 delay1 = mpt::saturate_round<uint32>(static_cast<float>(delay0) * 1.18920707f);  // 2^0.25
+	uint32 delay2 = mpt::saturate_round<uint32>(static_cast<float>(delay1) * 1.18920707f);
+	uint32 delay3 = mpt::saturate_round<uint32>(static_cast<float>(delay2) * 1.18920707f);
+	uint32 delay4 = mpt::saturate_round<uint32>(static_cast<float>(delay0 + delay2) * 0.11546667f);
+	uint32 delay5 = mpt::saturate_round<uint32>(static_cast<float>(delay1 + delay3) * 0.11546667f);
 	// Comb delays
 	m_delay[0] = delay0 - delay4;
 	m_delay[1] = delay2 - delay4;
@@ -226,7 +226,7 @@ CString WavesReverb::GetParamDisplay(PlugParamIndex param)
 void WavesReverb::RecalculateWavesReverbParams()
 {
 	// Recalculate filters
-	const double ReverbTimeSmp = -3000.0 / static_cast<double>(m_SndFile.GetSampleRate() * ReverbTime());
+	const double ReverbTimeSmp = -3000.0 / (static_cast<double>(m_SndFile.GetSampleRate()) * static_cast<double>(ReverbTime()));
 	const double ReverbTimeSmpHF = ReverbTimeSmp * (1.0 / static_cast<double>(HighFreqRTRatio()) - 1.0);
 
 	m_coeffs[0] = static_cast<float>(std::pow(10.0, m_delay[4] * ReverbTimeSmp));
